# api-docs.yaml
openapi: 3.0.1
info:
  title: iBoard 社區公佈欄(後端Spring boot)
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /facilities/{facilityId}:
    get:
      tags:
      - 公共設施租借 CRUD
      summary: 取得設施預定情況
      description: 透過facility_id取得設施預定資料
      operationId: getFacilityById
      parameters:
      - name: facilityId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Facility'
    put:
      tags:
      - 公共設施租借 CRUD
      summary: 修改預約
      description: 透過facility_id取得設施預定資料後修改
      operationId: updateFacility
      parameters:
      - name: facilityId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FacilityRequest'
        required: true
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Facility'
        "404":
          description: 無此 設施 ID，無法執行修改功能
    delete:
      tags:
      - 公共設施租借 CRUD
      summary: 刪除預約
      description: 透過facility_id刪除預約
      operationId: deleteFacility
      parameters:
      - name: facilityId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: 成功
...
  /facilities:
    get:
      tags:
      - 公共設施租借 CRUD
      summary: 取得設施預定列表
      description: 可模糊查詢姓名，可選擇場地，可選擇日期，可選擇時段，可以排除掉已過期資料(也可以全部都顯示)，依照前端請求orderBy指定項目排序(預設為預定日期時間)，依照前端請求sort指定排序順序(預設ASC由小到大排序)，
      operationId: getFacilities_1
      parameters:
      - name: name
        in: query
        description: 租借者
        required: false
        schema:
          type: string
      - name: category
        in: query
        description: 場地
        required: false
        schema:
          type: string
          enum:
          - GYM
          - MEETING_ROOM
          - TABLE_TENNIS
      - name: day
        in: query
        description: 租借日期
        required: false
        schema:
          type: string
      - name: time
        in: query
        description: 租借時間
        required: false
        schema:
          type: string
      - name: today
        in: query
        description: 可設定排除時間點
        required: false
        schema:
          type: string
      - name: orderBy
        in: query
        description: 根據所選欄位排序
        required: false
        schema:
          type: string
          default: reserved_day
      - name: sort
        in: query
        description: 排序方式
        required: false
        schema:
          type: string
          default: ASC
      responses:
        "404":
          description: 無此 RFID ID，無法執行修改功能
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Facility'
    post:
      tags:
      - 公共設施租借 CRUD
      summary: 新增預約設施
      description: 新增預約設施
      operationId: getFacilities
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FacilityRequest'
        required: true
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Facility'
...
components:
  schemas:
...
    FacilityRequest:
      required:
      - facilityCategory
      - reservedDay
      - reservedTime
      - userAddress
      - userName
      type: object
      properties:
        userAddress:
          type: string
          description: 預定者地址
        userName:
          type: string
          description: 預定者姓名
        facilityCategory:
          type: string
          description: 場地
          enum:
          - GYM
          - MEETING_ROOM
          - TABLE_TENNIS
        reservedDay:
          type: string
          description: 預定日期
          format: date
        reservedTime:
          type: string
          description: 預定時間
    Facility:
      type: object
      properties:
        facilityId:
          type: integer
          description: 公共設施 ID
          format: int32
        userAddress:
          type: string
          description: 租借人地址
        userName:
          type: string
          description: 租借人姓名
        facilityCategory:
          type: string
          description: 租借人租的設施
          enum:
          - GYM
          - MEETING_ROOM
          - TABLE_TENNIS
        reservedDay:
          type: string
          description: 預定的日期
          format: date
        reservedTime:
          type: string
          description: 預定的時間(ex:10:00-12:00)
        createdDate:
          type: string
          description: 創建預定數據時間
          format: date-time
        lastModifiedDate:
          type: string
          description: 最後修改預定數據時間
          format: date-time
...